requirements.txt
openai>=1.0.0
pydantic>=2.0.0
jinja2>=3.1.0
rich>=13.0.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
